+++ app/src/main/java/com/notes/application/MainActivity.kt
@@ -6,7 +6,7 @@ import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
-import com.notes.presentation.NoteAppNavigation
+import com.notes.presentation.ui.NoteAppNavigation
 import com.notes.presentation.theme.NotesTheme
 import dagger.hilt.android.AndroidEntryPoint
