+++ app/src/main/java/com/notes/data/database/NoteDatabase.kt
@@ -1,4 +1,4 @@
-package com.notes.data
+package com.notes.data.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
