+++ app/src/main/java/com/notes/presentation/ui/TopBar.kt
@@ -1,4 +1,4 @@
-package com.notes.presentation
+package com.notes.presentation.ui
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
